Add configurations cypress feature and steps
Steps configuration.
File .feature update tags
