+++ cypress/e2e/Login.feature
@@ -1,7 +1,9 @@
+#@LoginWebPage
 Feature: Test Login Page For Demo Cypress.io
 
     This is a test automation demo with Cypress, Cucumber and NodeJS tools and technologies.
 
+    @LoginWebPage @ValidCredential
     Scenario: User Should Be Able To Login Using Valid Credentials
         Given Visit Practice Test Automation Website
         Then User Provide Username
@@ -10,10 +12,10 @@ Feature: Test Login Page For Demo Cypress.io
         Then User Read The Text Of Successfully Log In Into The Website
         Then User Click On Logout Button For Reply To The Login Website
 
-        # Scenario: User Should Be Able To Login Using Invalid Credentials
-        # Given Visit Practice Test Automation Website
-        # Then User Provide Username
-        # Then User Provide Password
-        # Then User Click On Login Button To Log In Into The Practice Test Automation Website
-        # Then User Read The Text Of Successfully Log In Into The Website
-        # Then User Click On Logout Button For Reply To The Login Website
+# Scenario: User Should Be Able To Login Using Invalid Credentials
+# Given Visit Practice Test Automation Website
+# Then User Provide Username
+# Then User Provide Password
+# Then User Click On Login Button To Log In Into The Practice Test Automation Website
+# Then User Read The Text Of Successfully Log In Into The Website
+# Then User Click On Logout Button For Reply To The Login Website
